make backends folder, start docstrings
