+++ src/ixdat/db.py
@@ -1,158 +1,26 @@
 """This module contains the classes which pass on database functionality"""
 
-import json
-import numpy as np
-from .config import STANDARD_DATA_DIRECTORY
 from .exceptions import DataBaseError
+from .backends import DATABASE_BACKENDS
 
 
-def id_from_path(path):
-    try:
-        return int(path.stem.split("_")[0])
-    except ValueError:
-        return None
+MemoryBackend = DATABASE_BACKENDS["memory"]
+memory_backend = MemoryBackend()
+DirBackend = DATABASE_BACKENDS["directory"]
 
 
-def name_from_path(path):
-    return "".join(path.stem.split("_")[1:])
-
-
-class LocalBackend:
-    def __init__(self):
-        self.next_available_ids = {}
-
-    def get_next_available_id(self, table_name):
-        if table_name in self.next_available_ids:
-            i = self.next_available_ids[table_name]
-            self.next_available_ids[table_name] += 1
-        else:
-            i = 1
-            self.next_available_ids[table_name] = 2
-        return i
-
-
-local_backend = LocalBackend()
-
-
-class DirBackend:
-    """A database backend that loads and saves .ix files from a directory"""
-
-    def __init__(
-        self, directory=STANDARD_DATA_DIRECTORY, suffix="ix", data_suffix="ixdata"
-    ):
-        """Initialize a directory database backend with the directory as Path"""
-        self.directory = directory
-        self.suffix = suffix
-        self.data_suffix = data_suffix
-
-    @property
-    def name(self):
-        return f"DirBackend({self.directory})"
-
+class BackendBase:
     def save(self, obj):
-        if obj.data_objects:
-            # save any data objects first as this may change the references
-            for data_obj in obj.data_objects:
-                self.save_data_obj(data_obj)
-        try:
-            table_name = obj.table_name
-        except AttributeError:
-            table_name = str(type(obj))
-        obj_as_dict = obj.as_dict()
-        obj.id = self.add_row(obj_as_dict, table_name=table_name)
-        obj.backend_name = self.name
-        return obj.id
+        raise NotImplementedError
 
     def open(self, cls, i):
-        table_name = cls.table_name
-        obj_as_dict = self.open_serialization(table_name, i)
-        obj = cls.from_dict(obj_as_dict)
-        obj.backend = self
-        return obj
-
-    def contains(self, table_name, i):
-        return i in self.get_id_list(table_name)
+        raise NotImplementedError
 
     def load_obj_data(self, obj):
-        if not hasattr(obj, "data"):
-            # there's no data to be got or obj already has its data
-            return
-        path_to_row = self.get_path_to_row(obj.table_name, obj.id)
-        try:
-            return np.load(path_to_row.with_suffix(self.data_suffix))
-        except FileNotFoundError:
-            # there's no data to be got.
-            return
-
-    def save_data_obj(self, data_obj):
-        table_name = data_obj.table_name
-        if data_obj.backend == self and self.contains(table_name, data_obj.id):
-            return data_obj.id  # already saved!
-        obj_as_dict = data_obj.as_dict()
-        data = obj_as_dict["data"]
-        obj_as_dict["data"] = None
-        data_obj.id = self.add_row(obj_as_dict, table_name=table_name)
-        folder = self.directory / table_name
-        data_obj.backend = self
-        data_file_name = f"{data_obj.id}_{data_obj.name}.{self.data_suffix}"
-        np.save(folder / data_file_name, data)
-        return data_obj.id
-
-    def add_row(self, obj_as_dict, table_name):
-        folder = self.directory / table_name
-        if not folder.exists():
-            folder.mkdir()
-        i = self.get_next_available_id(table_name)
-
-        name = obj_as_dict["name"]
-        file_name = f"{id}_{name}.{self.suffix}"
-
-        with open(folder / file_name, "w") as f:
-            json.dump(obj_as_dict, f)
-        return i
-
-    def open_serialization(self, table_name, i):
-        path_to_row = self.get_path_to_row(table_name, i)
-        with open(path_to_row, "r") as file:
-            obj_as_dict = json.load(file)
-        return obj_as_dict
-
-    def get_path_to_row(self, table_name, i):
-        folder = self.directory / table_name
-        try:
-            return next(
-                p
-                for p in folder.iterdir()
-                if id_from_path(p) == i and p.suffix == self.suffix
-            )
-        except StopIteration:
-            return None
-
-    def get_id_list(self, table_name):
-        folder = self.directory / table_name
-        id_list = []
-        for file in folder.iterdir():
-            if not file.is_dir():
-                try:
-                    i = int(file.name.split("_")[0])
-                    id_list.append(i)
-                except TypeError:
-                    pass
-        return id_list
+        raise NotImplementedError
 
     def get_next_available_id(self, table_name):
-        return max(self.get_id_list(table_name)) + 1
-
-    def __eq__(self, other):
-        if other is self:
-            return True
-        if (
-            hasattr(other, "directory")
-            and other.directory.resolve() == self.directory.resolve()
-            and other.directory.lstat() == self.directory.lstat()
-        ):
-            return True
-        return False
+        raise NotImplementedError
 
 
 class DataBase:
@@ -201,7 +69,7 @@ class Saveable:
             setattr(self, attr, value)
         if self_as_dict and not self.column_attrs:
             self.column_attrs = {attr: attr for attr in self_as_dict.keys()}
-        self.backend_name = "local"  # SHOULD BE SET AFTER __INIT__ FOR LOADED OBJECT
+        self.backend_name = "memory"  # SHOULD BE SET AFTER __INIT__ FOR LOADED OBJECT
         self._id = None  # SHOULD BE SET AFTER THE __INIT__ OF INHERITING CLASSES
         self.name = None  # MUST BE SET IN THE __INIT__ OF INHERITING CLASSES
 
@@ -211,8 +79,8 @@ class Saveable:
     @property
     def id(self):
         if not self._id:
-            if self.backend_name == "local":
-                self._id = local_backend.get_next_available_id(self.table_name)
+            if self.backend_name == "memory":
+                self._id = memory_backend.get_next_available_id(self.table_name)
             else:
                 raise DataBaseError(
                     f"{self} comes from {self.backend_name} "
