+++ src/ixdat/techniques/ec_ms.py
@@ -1,6 +1,8 @@
+"""Module for representation and analysis of EC-MS measurements"""
+
 from .ec import ECMeasurement
 from .ms import MSMeasurement
 
 
 class ECMSMeasurement(ECMeasurement, MSMeasurement):
-    pass
+    """Class implementing raw EC-MS functionality"""
